pyproject.toml
[project]
name = "cardwatch-reporting-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "fpdf>=1.7.2",
    "openai>=1.58.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
